Adjust the proxyquire to work with TabularAPIserv
